:sparkles: add booking page (maybe working)
